+++ gittest/__init__.py
@@ -2,4 +2,8 @@
 A simple python lib
 """
 
-from gittest.osops import get_filename as get_filename, get_foldername as get_foldername
+from gittest.osops import (
+    get_filename as get_filename,
+    get_foldername as get_foldername,
+    join as join,
+)